refactor: Rename NextLinkAnchor to NextLink
