+++ src/components/atoms/NextLink/index.js
@@ -1,0 +1,1 @@
+export { default } from './NextLink';
