--- src/components/atoms/NextLinkAnchor/index.js
@@ -1,1 +1,0 @@
-export { default } from './NextLinkAnchor';
