+++ src/components/molecules/index.js
@@ -1,5 +1,3 @@
 export { default as AppBar } from './AppBar';
 export { default as Backtrack } from './Backtrack';
-export { default as DialogUploadImage } from './DialogUploadImage';
 export { default as FridgeInformation } from './FridgeInformation';
-export { default as FridgeDialog } from './FridgeDialog';
